requirements.txt
Flask==3.1.2
flask-cors==3.0.10
mysql-connector-python==8.1.0
deep-translator==1.9.7
python-dotenv==1.0.0
Werkzeug==2.3.0
gunicorn==21.2.0
